+++ BackEND/Spring/DrugDesributionManagementSystem/DrugDesributionManagementSystem/src/main/java/com/cdac_project/project/repository/DistributorRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cdac_project.project.model.Distributor;
 
-public interface DistributorRepository extends JpaRepository<Distributor, Long> {
+public interface DistributorRepository extends JpaRepository<Distributor, Integer> {
 
 	public Distributor findByEmail(String distributorEmail);
 }
