+++ BackEND/Spring/DrugDesributionManagementSystem/DrugDesributionManagementSystem/src/main/java/com/cdac_project/project/repository/MedicineCategoryRepository.java
@@ -8,11 +8,11 @@ import org.springframework.data.repository.query.Param;
 
 import com.cdac_project.project.model.MedicineCategory;
 
-public interface MedicineCategoryRepository extends JpaRepository<MedicineCategory, Long>{
+public interface MedicineCategoryRepository extends JpaRepository<MedicineCategory, Integer>{
 
-   	 public  MedicineCategory findByCategoryid(Long categoryId);
+   	 public  MedicineCategory findByCategoryid(int categoryId);
 	 
 //	 @Query("Select c from medicine_db c Where c.Category_id=:Category_id")
 //	 public MedicineCategory findById
-//	 (@Param("Category_id") Long Category_id);
+//	 (@Param("Category_id") int Category_id);
 }
