+++ BackEND/Spring/DrugDesributionManagementSystem/DrugDesributionManagementSystem/src/main/java/com/cdac_project/project/service/DistributorService.java
@@ -7,7 +7,7 @@ import com.cdac_project.project.model.Distributor;
 
 @Service
 public interface DistributorService{
-	public  Distributor findDistributorById(Long Distributor_ID)throws DistributorException;
+	public  Distributor findDistributorById(int Distributor_ID)throws DistributorException;
 	
 	public  Distributor findDistributorProfileByJwt(String jwt)throws DistributorException;
